# api-gateway/docs/openapi.yaml
openapi: 3.1.0
info:
  title: PayNoval Gateway API
  version: 1.0.0
  description: |
    Passerelle d’orchestration vers plusieurs microservices (providers).
    - Auth globale: Bearer
    - **Idempotency-Key** requis pour les opérations de création
    - **X-Request-Id** pour la traçabilité (corrélation logs)
    - Headers d’audit propagés vers les providers: x-internal-token, x-user-id, x-session-id
    - Certains endpoints publics (simulateurs & monitoring) ne nécessitent pas d’auth.
servers:
  - url: https://api.paynoval.com
    description: Production
  - url: https://sandbox.api.paynoval.com
    description: Sandbox
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    XRequestId:
      name: X-Request-Id
      in: header
      required: true
      description: UUID (corrélation et audit)
      schema: { type: string, format: uuid }
    ProviderQuery:
      name: provider
      in: query
      required: false
      description: Provider cible pour router la requête
      schema:
        type: string
        enum: [paynoval, stripe, bank, mobilemoney, visa_direct, visadirect, cashin, cashout, stripe2momo, flutterwave]
  schemas:
    ApiError:
      type: object
      properties:
        success: { type: boolean, example: false }
        error: { type: string, example: "Erreur interne provider" }
        code: { type: string, example: "invalid_request" }
        details:
          type: object
      additionalProperties: false
    Transaction:
      type: object
      description: Objet transaction retourné par les providers (forme générique)
      properties:
        id: { type: string, example: "tx_65f1..." }
        reference: { type: string, example: "GW-2025-000123" }
        provider: { type: string, example: "mobilemoney" }
        amount: { type: number, example: 50000 }
        currency: { type: string, example: "XOF" }
        status: { type: string, example: "pending" }
        createdAt: { type: string, format: date-time }
        meta:
          type: object
          description: Métadonnées spécifiques au provider
      additionalProperties: true
    TransactionList:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Transaction' }
        total: { type: integer, example: 42 }
      additionalProperties: false
...
    FeesSimulateRequest:
      type: object
      properties:
        provider:
          type: string
          enum: [paynoval, stripe, bank, mobilemoney, visa_direct, visadirect, cashin, cashout, stripe2momo, flutterwave]
        amount: { type: number, example: 100.0 }
        currency: { type: string, example: "CAD" }
        country: { type: string, example: "CI" }
        destination: { type: string, example: "paynoval" }
      required: [amount]
      additionalProperties: true
    FeesSimulateResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        provider: { type: string, example: "mobilemoney" }
        adminFee: { type: number, example: 0.45 }
        totalFee: { type: number, example: 1.25 }
        effectiveRate: { type: number, example: 590.0 }
        currency: { type: string, example: "CAD" }
      additionalProperties: true
    CommissionsSimulateRequest:
      type: object
      properties:
        amount: { type: number, example: 100.0 }
        currency: { type: string, example: "CAD" }
      required: [amount]
      additionalProperties: false
    CommissionsSimulateResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        commission: { type: number, example: 0.5 }
        currency: { type: string, example: "CAD" }
    ExchangeRateResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        from: { type: string, example: "CAD" }
        to: { type: string, example: "XOF" }
        rate: { type: number, example: 590.12 }
        updatedAt: { type: string, format: date-time }
paths:
...
  /api/v1/transactions:
    get:
      tags: [Transactions]
      summary: Lister les transactions (proxy vers provider)
      description: |
        Route le GET vers `GET {provider}/transactions`.
        **Auth requise.**
      parameters:
        - $ref: '#/components/parameters/ProviderQuery'
        - $ref: '#/components/parameters/XRequestId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionList' }
        '4XX':
          description: Erreur côté client
          content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
        '5XX':
          description: Erreur provider/serveur
          content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
...
  /api/v1/fees/simulate:
    post:
      tags: [Fees]
      security: []   # public
      summary: Simuler les frais
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FeesSimulateRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeesSimulateResponse' }
  /api/v1/commissions/simulate:
    post:
      tags: [Commissions]
      security: []   # public
      summary: Simuler la commission cagnotte
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CommissionsSimulateRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommissionsSimulateResponse' }
  /api/v1/exchange-rates/rate:
    get:
      tags: [ExchangeRates]
      security: []   # public
      summary: Obtenir un taux de change
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, example: "CAD" }
        - name: to
          in: query
          required: true
          schema: { type: string, example: "XOF" }
        - name: amount
          in: query
          required: false
          schema: { type: number, example: 100 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExchangeRateResponse' }

  # ================== MONITORING (PUBLICS) ==================
...